Task3_SignLanguageDetection/requirements.txt
tensorflow==2.5.0
numpy==1.19.5
opencv-python==4.5.3.56
Pillow==8.3.1